feat(csharp): add consumer and publisher clients (#2259)
- Add IggyPublisher for sending messages:
  - Create stream and topic
  - Send direct or in background
  - Generic version with serializer
- Add IggyConsumer for polling messages:
  - Create and join consumer group
  - Store offset support
  - IAsyncEnumerable support
  - Generic version with deserializer
- Add builders for both clients
- Implement AES encryptor
- Change `Message` from readonly struct to class

---------

+++ foreign/csharp/Iggy_SDK_Tests/Utils/Messages/MessageFactory.cs
@@ -226,7 +226,7 @@ internal static class MessageFactory
     {
         return new MessageFetchRequest
         {
-            Count = Random.Shared.Next(1, 10),
+            Count = (uint)Random.Shared.Next(1, 10),
             AutoCommit = true,
             Consumer = Consumer.New(1),
             PartitionId = (uint)Random.Shared.Next(1, 10),
@@ -241,7 +241,7 @@ internal static class MessageFactory
     {
         return new MessageFetchRequest
         {
-            Count = count,
+            Count = (uint)count,
             AutoCommit = true,
             Consumer = Consumer.New(consumerId),
             PartitionId = partitionId,
@@ -256,7 +256,7 @@ internal static class MessageFactory
     {
         return new MessageFetchRequest
         {
-            Count = count,
+            Count = (uint)count,
             AutoCommit = true,
             Consumer = Consumer.Group(consumerGroupId),
             PartitionId = partitionId,
